Fixed measurements showing
Moved the measurements out from the frame, and in front of the image, so it can be seen.
Changed a little bit with the scale so xy are tied to each.

Line 61: Added SetText() into the placing of normal objects
Line 76: Tagging custom frames as Placable
Line 78: Replaced old arguments with new to work off of the object's scale, not the value)
Line 85: As above
Line 98 - Line111: Changing the position.